Continuation of complete PC implementation.
CrossRoads fully operational, some automatized generator should be implemented.
I'm too destroyed to give any more sophisticated comments at the moment...

+++ core/src/main/java/paramatricCurves/FunctionT.java
@@ -8,7 +8,7 @@ public class FunctionT {
 		this.def = def;
 	}
 	
-	public double getFt(double t) {
+	public float getFt(float t) {
 		return def.getFt(t);
 	}
