+++ core/src/main/java/paramatricCurves/Get_F_t.java
@@ -2,5 +2,5 @@ package paramatricCurves;
 
 public interface Get_F_t {
 
-	public double getFt(double t);
+	public float getFt(float t);
 }
